Don't delete old options file unless new one was written

+++ src/utils/os_utils.py
@@ -28,7 +28,11 @@ def migrate_from_file(old_path: str, new_path: str, fallback: str):
         else:
             write_to_path(new_path, fallback)
 
-    if os.path.exists(old_path) and old_path != new_path:
+    if (
+        os.path.exists(old_path)
+        and os.path.exists(new_path)
+        and old_path != new_path
+    ):
         try:
             os.remove(old_path)
         except:
